add aarch64 port, reliable build system, and agent documentation
- Restore aarch64 port files from deleted branch (mkfile-Linux-aarch64,
  emu/Linux/{asm,segflush,tas}-aarch64, lib9 aarch64 asm, libinterp
  comp/das-aarch64.c, Linux/aarch64/include headers, libkern mkfile)
- Fix aarch64 portability: uintptr_t casts in lib9/vfprint/sprint/fmtfd,
  mkstemp in utils/mk/Posix.c, stdint.h includes, time.h in mk
- Add Makefile: reliable fresh-build wrapper that always nukes before
  rebuilding each component in EMUDIRS order, bypassing mk's unreliable
  dependency tracking
- Add AGENTS.md: top-level agent onboarding (repo layout, build system,
  key data structures, coding conventions, common tasks)
- Add AGENTS_DIS.md: Dis VM reference (instruction set, .dis file format,
  interpreter loop, GC, channels, alt statement, JIT)
- Add AGENTS_LIMBO.md: Limbo language and compiler reference (syntax,
  type system, module system, compiler pipeline, standard modules)
- Add AGENTS_KERNEL.md: kernel internals reference (9P protocol, device
  drivers, Chan, namespace/mounts, scheduler, memory pools, OS layer)
- Add AGENTS_AARCH64.md: AArch64 architecture reference for VM porting
  (register file, calling convention, ISA, memory model, atomics, cache
  coherency for JIT, signal handling, dispatch techniques)

+++ lib9/sprint.c
@@ -26,7 +26,7 @@ sprint(char *buf, char *fmt, ...)
 	 * we must be sure not to overflow a 32-bit pointer.
 	 */
 	if((uintptr)buf+len < (uintptr)buf)
-		len = -(uint)buf-1;
+		len = -(uint)(uintptr)buf-1;
 
 	va_start(args, fmt);
 	n = vsnprint(buf, len, fmt, args);
