+++ utils/mk/archive.c
@@ -1,4 +1,5 @@
 #include	"mk.h"
+#include	<time.h>
 #include	<ar.h>
 
 static int dolong;
@@ -22,13 +23,13 @@ atimeof(int force, char *name)
 	if(sym){
 		if(force || (t > (long)sym->value)){
 			atimes(archive);
-			sym->value = (void *)t;
+			sym->value = (void *)(uintptr_t)t;
 		}
 	}
 	else{
 		atimes(archive);
 		/* mark the aggegate as having been done */
-		symlook(strdup(archive), S_AGG, "")->value = (void *)t;
+		symlook(strdup(archive), S_AGG, "")->value = (void *)(uintptr_t)t;
 	}
 		/* truncate long member name to sizeof of name field in archive header */
 	if(dolong)
@@ -37,7 +38,7 @@ atimeof(int force, char *name)
 		snprint(buf, sizeof(buf), "%s(%.*s)", archive, SARNAME, member);
 	sym = symlook(buf, S_TIME, 0);
 	if (sym)
-		return (long)sym->value;	/* uggh */
+		return (long)(uintptr_t)sym->value;	/* uggh */
 	return 0;
 }
 
@@ -116,7 +117,7 @@ atimes(char *ar)
 			continue;
 		}
 		sprint(buf, "%s(%s)", ar, n);
-		symlook(strdup(buf), S_TIME, (void *)t)->value = (void *)t;
+		symlook(strdup(buf), S_TIME, (void *)(uintptr_t)t)->value = (void *)(uintptr_t)t;
 		t = atol(h.size);
 		if(t&01) t++;
 		LSEEK(fd, t, 1);
